+++ tests/Weblog/application/models/Comment.php
@@ -5,13 +5,6 @@ class Comment extends Rougin\Wildfire\Model
     use Rougin\Wildfire\Traits\ValidateTrait;
 
     /**
-     * The attributes that should be visible for serialization.
-     *
-     * @var array
-     */
-    protected $hidden = array('id');
-
-    /**
      * An array of validation rules. This needs to be the same format
      * as validation rules passed to the Form_validation library.
      *